requirements.txt
fastapi
pydantic
uvicorn
dotenv

tensorflow
keras
numpy
pillow
